wrapping up changes for AudioPlayer tests
